requirements.txt
fastapi
uvicorn[standard]
email-validator
sqlalchemy
aiosqlite
asyncpg
alembic
pydantic
python-dotenv
pytest
httpx
pytest-asyncio
black
isort